some refactoring to comply standards
