+++ src/User/UseCase/ToggleActivator.php
@@ -1,10 +1,10 @@
 <?php
 namespace User\UseCase;
 
-use User\Object\ToggleActivatorRequest;
+use User\UseCase\ToggleActivator\Request;
 
 interface ToggleActivator
 {
-    public function execute( ToggleActivatorRequest $request, ToggleActivatorResponseHandler $responseHandler );
+    public function execute( Request $request, ToggleActivatorResponseHandler $responseHandler );
 }
 //EOF ToggleActivator.php
